The input for entering the release was limited, now users can only select the release tag (release, hotfix, unscheduled) by clicking the button (in the window-with-input component). Manually, the user can now enter only the release number. To avoid input errors, the numeric-input directive was added, which allows you to enter only numbers in the input field. The TAGS constant has been added to the list of release tags. To output tags with a number from the backend, the approach to data transmission was changed, data-model was added, from where the ReleaseInfo interface is exported to describe the contents of the releaseList array from the ReleaseDataResponse interface.
Now users can add a release.
In the server request models, the AddingReleaseInfo interface has been added to transmit information about the tag entered by the user and the release number to the server. In the server response model, the DeleteReleaseResponse interface has been replaced with the universal ModifyReleaseListResponse for responses to both deletion and addition of a release. A request to add a release has been added to the data-service.

Added the ability to quickly select the release week: the left/right arrows select the month, and the up/down arrows select the week. Also, by pressing the Tab key, you can select the tag button, input and the release add button